Add filesystem abstraction, this does quite some things: * On UNIX/Linux, multiple data directories are now supported. Also spring correctly   recognizes read-only and read-write data directories. There must always be at least   one read-write data directory at this moment. * Almost no boost::filesystem code anymore. Once FileSystemHandler::FindFiles has been   implemented for Linux & Windows without boost we have one dependency less. * The functionality used from boost::fs is now provided by this FileSystem. * All functions reading/writing files should use filesystem.{fopen,ostream,istream},   or query a vector with filesystem.GetNativeFilenames() and try to do whatever they   need to do on all filenames in that vector (UNIX/Linux platforms support multiple   data directories now.)   I have tried to convert all of them but may have missed some. If you notice direct   filesystem access without using FileSystem functions, rewrite it and/or poke me. * No more big chunks of Linux only code in Main.cpp * Fixed several minor bugs here and there while I stumbled upon them. * I hope I didn't forget anything here.
Also applied (modified) patch by TheBlasphemer, fixing a bug regarding EOF detection
in hpiutil2, and removing boost::fs stuff from the same file because it throws on
windows.



git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1775 37977431-3df6-0310-b722-df95706aa16b

+++ rts/lib/hpiutil2/scrambledfile.h
@@ -56,7 +56,7 @@ namespace hpiutil
 		scrambledfile(const char *fname);
 		scrambledfile(std::string const &fname);
 		~scrambledfile();
-		boost::uint8_t read();
+		boost::uint32_t read();
 		boost::uint32_t read(boost::uint8_t *buf);
 		boost::uint32_t read(boost::uint8_t *buf,const boost::uint32_t off, const boost::uint32_t len);
 		boost::uint32_t readint();
